Add files via upload
It is possible to search in the data using multiple criteria at once. This search is for two separate groups of information. Within those two groups, the query seeks to find one person common to either group.

This is analogous to "(choir OR cheer) And Amanda" to find a musical, sports oriented student named Amanda, rather than the viola playing, snarky goth Amanda.