Add files via upload
It is possible to search in Splunk by asking for one thing, and definitely not another. In this case, the search query is for users, but not administrators. 

An analogous search on a company's database would be for "permanent employees NOT contractors" in the database of people working forty hours per week at a location.